requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.2
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.11
djangorestframework==3.16.1
gevent==25.9.1
greenlet==3.3.1
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
pycparser==3.0
python-dateutil==2.9.0.post0
redis==7.1.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
zope.event==6.1
zope.interface==8.2
